requirements.txt
cryptography>=40.0.0
tkinter
pillow>=9.4.0